refactor(editor): Move useClipboard into @n8n/composables behind re-export shim (no-changelog)
Relocate useClipboard (+ its test) from editor-ui into @n8n/composables via
subpath export (@n8n/composables/useClipboard). Fold in its only sink,
PopOutWindowKey, as a sibling injection-key module. The old @/app paths become
thin deprecated re-export shims so call sites are untouched.
