+++ packages/frontend/@n8n/composables/src/PopOutWindowKey.ts
@@ -1,0 +1,8 @@
+import type { InjectionKey, Ref } from 'vue';
+
+/**
+ * Injection key for an optional pop-out window. Consumers `provide` it from the
+ * host that owns the detached window; `useClipboard` reads it so a copy issued
+ * inside the pop-out targets that window's `navigator.clipboard`.
+ */
+export const PopOutWindowKey: InjectionKey<Ref<Window | undefined>> = Symbol('PopOutWindow');
